# docs/swagger.yaml
openapi: 3.0.3
info:
  title: MacBook Back Template API
  version: 1.0.0
  description: >
    Exported API contract from current NestJS controllers. All responses are wrapped
    by a global response interceptor as `{ success, statusCode, body: { data, message, type? } }`.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /units:
    get:
      tags: [UnitsHierarchy]
      summary: Fetch all active units with direct parent
      description: Returns flattened list of active units with status and immediate parent details.
      parameters:
        - $ref: '#/components/parameters/ScreenDateHeader'
        - $ref: '#/components/parameters/UserHeader'
      responses:
        '200':
          description: Units fetched.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitsHierarchyResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /units/lower-levels:
    get:
      tags: [UnitsHierarchy]
      summary: Fetch lower-level units
      description: Returns up to ten lower-level units for the screen unit with hierarchy-style unit fields.
      parameters:
        - $ref: '#/components/parameters/ScreenDateHeader'
        - $ref: '#/components/parameters/UserHeader'
        - $ref: '#/components/parameters/OptionalUnitHeader'
        - name: filter
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 10
            default: 10
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: isConnectedToRoot
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Lower-level units fetched.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitsHierarchyResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /units/hierarchy:
    get:
      tags: [UnitsHierarchy]
      summary: Fetch active unit hierarchy graph
      description: Returns all active units for the screen date with parent, status, emergency, and `isConnectedToRoot` markers relative to the authenticated user's root unit.
      parameters:
        - $ref: '#/components/parameters/ScreenDateHeader'
        - $ref: '#/components/parameters/UserHeader'
        - $ref: '#/components/parameters/OptionalUnitHeader'
      responses:
        '200':
          description: Hierarchy fetched.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitsHierarchyResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'

    post:
      tags: [UnitsHierarchy]
      summary: Add hierarchy relation
      description: Adds parent-child relation (`upperUnit` -> `lowerUnit`) for current screen date.
      parameters:
        - $ref: '#/components/parameters/ScreenDateHeader'
        - $ref: '#/components/parameters/UserHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HierarchyRelationRequest'
      responses:
        '200':
          description: Relation added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageOnlyResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'

    delete:
      tags: [UnitsHierarchy]
      summary: Remove hierarchy relation
      description: Removes parent-child relation for current screen date.
      parameters:
        - $ref: '#/components/parameters/ScreenDateHeader'
        - $ref: '#/components/parameters/UserHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HierarchyRelationRequest'
      responses:
        '200':
          description: Relation removed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageOnlyResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'

    put:
      tags: [UnitsHierarchy]
      summary: Transfer hierarchy relation
      description: Moves `lowerUnit` from existing parent to new `upperUnit` for current screen date.
      parameters:
        - $ref: '#/components/parameters/ScreenDateHeader'
        - $ref: '#/components/parameters/UserHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HierarchyRelationRequest'
      responses:
        '200':
          description: Relation transferred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageOnlyResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
components:
  parameters:
    ScreenDateHeader:
      in: header
      name: screendate
      required: true
      description: Screen date in `YYYY-MM-DD`; required by global middleware.
      schema:
        type: string
        example: '2026-03-02'
    UserHeader:
      in: header
      name: user
      required: false
      description: Username for auditing; defaults to `S9107544` when missing.
      schema:
        type: string
        example: S1234567
...
    OptionalUnitHeader:
      in: header
      name: unit
      required: false
      description: Optional root unit id (defaults to `1` for this endpoint).
      schema:
        type: integer
        example: 1234
  responses:
    ErrorResponse:
      description: Error envelope returned by global exception filter.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    MessageType:
      type: string
      enum: [Success, Failure, Fatal]
    ErrorResponse:
      type: object
      required: [success, statusCode, body]
      properties:
        success:
          type: boolean
          enum: [false]
        statusCode:
          type: integer
          example: 400
        body:
          type: object
          required: [data, message]
          properties:
            data:
              nullable: true
              default: null
            message:
              type: string
            type:
              $ref: '#/components/schemas/MessageType'
...
    UnitStatus:
      type: object
      required: [id, description]
      properties:
        id:
          type: integer
        description:
          type: string
    UnitParent:
      type: object
      required: [id, description, level, simul, status]
      properties:
        id:
          type: integer
        description:
          type: string
        level:
          type: integer
        simul:
          type: string
        status:
          $ref: '#/components/schemas/UnitStatus'
    UnitNode:
      type: object
      description: Hierarchy unit entity.
      required: [id, description, level, simul, isEmergencyUnit, status, parent]
      properties:
        id:
          type: integer
        description:
          type: string
        level:
          type: integer
        simul:
          type: string
        isEmergencyUnit:
          type: boolean
        isConnectedToRoot:
          type: boolean
        isRootUnit:
          type: boolean
        status:
          $ref: '#/components/schemas/UnitStatus'
        parent:
          allOf:
            - $ref: '#/components/schemas/UnitParent'
          nullable: true
...
    HierarchyRelationRequest:
      type: object
      description: Entity payload for add/remove/transfer hierarchy operations.
      required: [lowerUnit, upperUnit, rootUnit]
      properties:
        lowerUnit:
          type: integer
        upperUnit:
          type: integer
        rootUnit:
          type: integer
          nullable: true
...
    MessageOnlyResponse:
      type: object
      required: [success, statusCode, body]
      properties:
        success:
          type: boolean
          enum: [true]
        statusCode:
          type: integer
        body:
          type: object
          required: [message]
          properties:
            data:
              nullable: true
              default: null
            message:
              type: string
            type:
              $ref: '#/components/schemas/MessageType'
...
    UnitsHierarchyResponse:
      type: object
      required: [success, statusCode, body]
      properties:
        success:
          type: boolean
          enum: [true]
        statusCode:
          type: integer
        body:
          type: object
          required: [data, message]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/UnitNode'
            message:
              type: string
            type:
              $ref: '#/components/schemas/MessageType'